requirements.txt
rembg==2.0.29
replicate==0.4.0
opencv-python-headless==4.6.0.66
google-cloud-storage==2.7.0
fastapi==0.87.0
uvicorn==0.20.0
pydantic==1.9.2
numpy==1.23.5